+++ src/main.cpp
@@ -6,7 +6,8 @@
 
 #include "CLI/CLI.hpp"
 
-#include "fugle_client.hpp"
+#include "fugle_api_list.hpp"
+#include "fugle_client_base.hpp"
 
 using namespace std;
 using namespace fugle_realtime;
@@ -15,7 +16,15 @@ struct Args {
   string filePath = "api_key.txt";
   string endpoint = "/intraday/quote";
   string symbol = "2330";
+  bool showApiList = false;
 };
+
+int ShowAPIs() {
+  for (const auto &s : kFugleAPIList) {
+    spdlog::info(s);
+  }
+  return 0;
+}
 
 int main(int argc, char **argv) {
   CLI::App app{"Fugle Terminal"};
@@ -24,6 +33,7 @@ int main(int argc, char **argv) {
   app.add_option("-f,--api-file", args.filePath, "Path to api key");
   app.add_option("-s,--stock-symbol", args.symbol, "symbol");
   app.add_option("-e,--endpoint", args.endpoint, "endpoint");
+  app.add_flag("-l,--show-api-list", args.showApiList, "print Fugle APIs");
 
   try {
     app.parse(argc, argv);
@@ -31,14 +41,17 @@ int main(int argc, char **argv) {
     return app.exit(e);
   }
 
-  std::ifstream apiKeyFile(args.filePath);
+  if (args.showApiList) {
+    return ShowAPIs();
+  }
+
+  std ::ifstream apiKeyFile(args.filePath);
   if (!apiKeyFile.is_open()) {
     spdlog::error("Can not open the api key file");
     return 1;
   }
   string apiKey;
   getline(apiKeyFile, apiKey);
-  spdlog::debug("Key {}", apiKey);
 
   FugleHttpClientBase fugleClient(apiKey);
